refactor: dedupe wire-shape helpers into internal/wireshape (PR self-review)
Self-review of PR #130 surfaced three real issues:

1. Seven helper functions were duplicated across the test file and
   the refresh script (schemasFromSpecs, discoverSDKTypes,
   extractWireFieldsFromAST, parseJSONTagName, yamlNodeToMap,
   yamlNodeToValue, difference). A future change to one copy would
   silently let the gate and the refresher drift apart.

   Fix: extracted all of them into internal/wireshape. The SDK
   package's wire-shape test imports it, and the refresh command
   under scripts/... imports it. Consumers of the SDK cannot import
   internal/... packages, so the yaml.v3 dep stays out of any
   consumer binary. Eliminates the drift risk at the language level.

2. The refresh script wrote the baseline non-atomically:
   os.Create + enc.Encode would leave a truncated file behind if
   encoding failed partway through, which the very next CI run
   would fail on at JSON parse.

   Fix: writeBaselineAtomically writes to a sibling temp file and
   os.Renames on success. Failed encode cleans up the temp file
   and never touches the real baseline.

3. No unit test locked the "yaml last-wins on duplicate keys
   matches Python's yaml.safe_load" claim made in the LoadSchemas
   comment. A refactor to yaml.Unmarshal into map[string]any
   would silently regress the claim (yaml.v3 errors on duplicates).

   Fix: internal/wireshape/wireshape_test.go now has
   TestNodeToMapLastWinsOnDuplicateKey plus tests for the positive
   and negative cross-spec-duplicate cases and ParseJSONTagName
   edge cases.

Verified:
- `go build ./...` clean
- `go vet ./...` clean
- `gofmt -s -l .` empty
- 5 internal/wireshape tests green
- 7 TestWireShape* tests at package root green against pinned specs
- Full `go test ./...` suite green
- Refresh script still produces identical baseline (git diff empty
  after regen), no .tmp.* files left on disk

LOC delta: 554 insertions / 564 deletions. The reshuffle ended up
slightly net-smaller after eliminating duplicated helpers.
