+++ scripts/refresh_wire_shape_baseline/main.go
@@ -5,15 +5,16 @@
 //
 //	go run ./scripts/refresh_wire_shape_baseline <specs_dir> [--sha <SHA>]
 //
-// When --sha is not provided, the script tries `git -C <parent-of-specs_dir>
+// When --sha is not provided, the script tries `git -C <specs_dir>
 // rev-parse HEAD` to pick up the commit of the specs repo. If neither
 // is available, the script exits non-zero rather than emit a poisoned
-// baseline with an empty openapi_specs_sha — the next CI run would fail
-// at workflow bootstrap anyway, fail fast here.
+// baseline with an empty openapi_specs_sha — the next CI run would
+// fail at workflow bootstrap anyway, so we fail fast here.
 //
-// The script uses the same loader functions as contract_wire_shape_test.go
-// via the axonflow package (tests and this command live in the same
-// module), so the extraction logic is not duplicated.
+// All extraction logic is shared with contract_wire_shape_test.go via
+// internal/wireshape so this tool and the gate cannot silently drift
+// apart. The output file is written via temp-file + rename so a
+// mid-encode failure never leaves a truncated baseline behind.
 
 package main
 
@@ -21,33 +22,16 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
-	"go/ast"
-	"go/parser"
-	"go/token"
 	"os"
 	"os/exec"
 	"path/filepath"
-	"reflect"
 	"sort"
 	"strings"
 
-	yaml "gopkg.in/yaml.v3"
+	"github.com/getaxonflow/axonflow-sdk-go/v5/internal/wireshape"
 )
 
 const baselineOut = "testdata/wire_shape_baseline.json"
-
-type baseline struct {
-	Comment             string                         `json:"_comment"`
-	OpenAPISpecsSHA     string                         `json:"openapi_specs_sha"`
-	CrossSpecDuplicates map[string]map[string][]string `json:"cross_spec_duplicates"`
-	RegisteredTypes     []string                       `json:"registered_types"`
-	PerTypeDrift        map[string]driftEntry          `json:"per_type_drift"`
-}
-
-type driftEntry struct {
-	SDKOnly  []string `json:"sdk_only"`
-	SpecOnly []string `json:"spec_only"`
-}
 
 func main() {
 	shaFlag := flag.String("sha", "", "community-repo commit SHA to pin")
@@ -65,6 +49,8 @@ func main() {
 		os.Exit(2)
 	}
 
+	// Resolve SHA before any expensive work so a bad run terminates
+	// early without touching the baseline file.
 	sha := *shaFlag
 	if sha == "" {
 		sha = gitHeadSHA(specsDir)
@@ -80,37 +66,35 @@ func main() {
 		os.Exit(2)
 	}
 
-	merged, duplicates, err := schemasFromSpecs(specsDir)
+	merged, duplicates, err := wireshape.LoadSchemas(specsDir)
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "error: %v\n", err)
 		os.Exit(1)
 	}
-	// Resolve SDK package dir relative to this tool's cwd. The refresher
-	// is run from the SDK repo root, so "." is the right pkg dir.
-	sdk, err := discoverSDKTypes(".")
+	sdk, err := wireshape.DiscoverSDKTypes(".")
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "error: %v\n", err)
 		os.Exit(1)
 	}
 
 	registered := []string{}
-	drift := map[string]driftEntry{}
+	drift := map[string]wireshape.DriftEntry{}
 	for name, sdkFields := range sdk {
 		specFields, ok := merged[name]
 		if !ok {
 			continue
 		}
 		registered = append(registered, name)
-		sdkOnly := difference(sdkFields, specFields)
-		specOnly := difference(specFields, sdkFields)
+		sdkOnly := wireshape.Difference(sdkFields, specFields)
+		specOnly := wireshape.Difference(specFields, sdkFields)
 		if len(sdkOnly) == 0 && len(specOnly) == 0 {
 			continue
 		}
-		drift[name] = driftEntry{SDKOnly: sdkOnly, SpecOnly: specOnly}
+		drift[name] = wireshape.DriftEntry{SDKOnly: sdkOnly, SpecOnly: specOnly}
 	}
 	sort.Strings(registered)
 
-	out := baseline{
+	out := wireshape.Baseline{
 		Comment: "Baseline of KNOWN wire-shape drift between the Go SDK and the " +
 			"OpenAPI specs. Generated by scripts/refresh_wire_shape_baseline. The CI " +
 			"gate fails on drift OUTSIDE this baseline. Entries here should be burned " +
@@ -122,23 +106,10 @@ func main() {
 		PerTypeDrift:        drift,
 	}
 
-	if err := os.MkdirAll(filepath.Dir(baselineOut), 0o755); err != nil {
+	if err := writeBaselineAtomically(baselineOut, &out); err != nil {
 		fmt.Fprintf(os.Stderr, "error: %v\n", err)
 		os.Exit(1)
 	}
-	f, err := os.Create(baselineOut)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "error: %v\n", err)
-		os.Exit(1)
-	}
-	enc := json.NewEncoder(f)
-	enc.SetIndent("", "  ")
-	if err := enc.Encode(&out); err != nil {
-		_ = f.Close()
-		fmt.Fprintf(os.Stderr, "error: %v\n", err)
-		os.Exit(1)
-	}
-	_ = f.Close()
 
 	fmt.Printf("Wrote baseline: %s\n", baselineOut)
 	fmt.Printf("  openapi_specs_sha:     %s\n", sha)
@@ -147,231 +118,50 @@ func main() {
 	fmt.Printf("  per_type_drift:        %d\n", len(drift))
 }
 
+// writeBaselineAtomically writes b to path via a sibling temp file +
+// rename, so a partial/interrupted encode cannot leave behind a
+// truncated baseline that would fail the CI bootstrap step.
+func writeBaselineAtomically(path string, b *wireshape.Baseline) error {
+	if err := os.MkdirAll(filepath.Dir(path), 0o755); err != nil {
+		return err
+	}
+	f, err := os.CreateTemp(filepath.Dir(path), filepath.Base(path)+".tmp.*")
+	if err != nil {
+		return err
+	}
+	tmpPath := f.Name()
+	cleanup := func() { _ = os.Remove(tmpPath) }
+	enc := json.NewEncoder(f)
+	enc.SetIndent("", "  ")
+	if err := enc.Encode(b); err != nil {
+		_ = f.Close()
+		cleanup()
+		return fmt.Errorf("encode baseline: %w", err)
+	}
+	if err := f.Close(); err != nil {
+		cleanup()
+		return fmt.Errorf("close temp file: %w", err)
+	}
+	if err := os.Rename(tmpPath, path); err != nil {
+		cleanup()
+		return fmt.Errorf("rename temp to %s: %w", path, err)
+	}
+	return nil
+}
+
+// gitHeadSHA reads the commit SHA of the git repo containing specDir,
+// or returns "" if the path isn't a git checkout or git isn't on PATH.
+// Best-effort: not a security boundary — the caller can always pass
+// --sha explicitly.
 func gitHeadSHA(specDir string) string {
 	git, err := exec.LookPath("git")
 	if err != nil {
 		return ""
 	}
-	cmd := exec.Command(git, "-C", specDir, "rev-parse", "HEAD")
+	cmd := exec.Command(git, "-C", specDir, "rev-parse", "HEAD") //nolint:gosec // fixed args, caller-supplied dir
 	out, err := cmd.Output()
 	if err != nil {
 		return ""
 	}
 	return strings.TrimSpace(string(out))
-}
-
-// ----- schema and SDK-type extractors, duplicated minimally from the
-// test file so this command has no dependency on the test build. -----
-
-func schemasFromSpecs(specDir string) (map[string][]string, map[string]map[string][]string, error) {
-	merged := map[string][]string{}
-	allDecls := map[string]map[string][]string{}
-
-	entries, err := os.ReadDir(specDir)
-	if err != nil {
-		return nil, nil, fmt.Errorf("read %s: %w", specDir, err)
-	}
-	names := make([]string, 0, len(entries))
-	for _, e := range entries {
-		if !strings.HasSuffix(e.Name(), ".yaml") {
-			continue
-		}
-		names = append(names, e.Name())
-	}
-	sort.Strings(names)
-
-	for _, name := range names {
-		full := filepath.Join(specDir, name)
-		data, err := os.ReadFile(full)
-		if err != nil {
-			return nil, nil, fmt.Errorf("read %s: %w", full, err)
-		}
-		var root yaml.Node
-		if err := yaml.Unmarshal(data, &root); err != nil {
-			return nil, nil, fmt.Errorf("parse %s: %w", full, err)
-		}
-		doc, ok := yamlNodeToMap(&root)
-		if !ok {
-			continue
-		}
-		comps, _ := doc["components"].(map[string]any)
-		if comps == nil {
-			continue
-		}
-		schemas, _ := comps["schemas"].(map[string]any)
-		for schemaName, raw := range schemas {
-			schema, ok := raw.(map[string]any)
-			if !ok {
-				continue
-			}
-			props, ok := schema["properties"].(map[string]any)
-			if !ok || len(props) == 0 {
-				continue
-			}
-			fields := make([]string, 0, len(props))
-			for k := range props {
-				fields = append(fields, k)
-			}
-			sort.Strings(fields)
-			if _, seen := allDecls[schemaName]; !seen {
-				allDecls[schemaName] = map[string][]string{}
-			}
-			allDecls[schemaName][name] = fields
-			merged[schemaName] = fields
-		}
-	}
-
-	duplicates := map[string]map[string][]string{}
-	for schemaName, decls := range allDecls {
-		if len(decls) < 2 {
-			continue
-		}
-		shapes := map[string]struct{}{}
-		for _, f := range decls {
-			shapes[strings.Join(f, "|")] = struct{}{}
-		}
-		if len(shapes) > 1 {
-			duplicates[schemaName] = decls
-		}
-	}
-	return merged, duplicates, nil
-}
-
-func discoverSDKTypes(pkgDir string) (map[string][]string, error) {
-	fset := token.NewFileSet()
-	pkgs, err := parser.ParseDir(fset, pkgDir, func(info os.FileInfo) bool {
-		return !strings.HasSuffix(info.Name(), "_test.go")
-	}, parser.ParseComments)
-	if err != nil {
-		return nil, fmt.Errorf("parse %s: %w", pkgDir, err)
-	}
-	result := map[string][]string{}
-	for _, pkg := range pkgs {
-		for _, file := range pkg.Files {
-			ast.Inspect(file, func(n ast.Node) bool {
-				ts, ok := n.(*ast.TypeSpec)
-				if !ok {
-					return true
-				}
-				if !ts.Name.IsExported() {
-					return true
-				}
-				st, ok := ts.Type.(*ast.StructType)
-				if !ok {
-					return true
-				}
-				fields := extractWireFieldsFromAST(st)
-				if len(fields) == 0 {
-					return true
-				}
-				result[ts.Name.Name] = fields
-				return true
-			})
-		}
-	}
-	return result, nil
-}
-
-func extractWireFieldsFromAST(st *ast.StructType) []string {
-	if st.Fields == nil {
-		return nil
-	}
-	out := []string{}
-	for _, field := range st.Fields.List {
-		if len(field.Names) == 0 {
-			continue
-		}
-		tagStr := ""
-		if field.Tag != nil {
-			tagStr = reflect.StructTag(strings.Trim(field.Tag.Value, "`")).Get("json")
-		}
-		wireName := parseJSONTagName(tagStr)
-		if wireName == "-" {
-			continue
-		}
-		for _, name := range field.Names {
-			if !name.IsExported() {
-				continue
-			}
-			if wireName != "" {
-				out = append(out, wireName)
-			} else {
-				out = append(out, name.Name)
-			}
-		}
-	}
-	sort.Strings(out)
-	return out
-}
-
-func parseJSONTagName(tag string) string {
-	if tag == "" {
-		return ""
-	}
-	if i := strings.Index(tag, ","); i >= 0 {
-		return tag[:i]
-	}
-	return tag
-}
-
-func difference(a, b []string) []string {
-	bs := map[string]struct{}{}
-	for _, s := range b {
-		bs[s] = struct{}{}
-	}
-	out := []string{}
-	for _, s := range a {
-		if _, ok := bs[s]; !ok {
-			out = append(out, s)
-		}
-	}
-	sort.Strings(out)
-	return out
-}
-
-func yamlNodeToMap(n *yaml.Node) (map[string]any, bool) {
-	v := yamlNodeToValue(n)
-	m, ok := v.(map[string]any)
-	return m, ok
-}
-
-func yamlNodeToValue(n *yaml.Node) any {
-	if n == nil {
-		return nil
-	}
-	switch n.Kind {
-	case yaml.DocumentNode:
-		if len(n.Content) == 0 {
-			return nil
-		}
-		return yamlNodeToValue(n.Content[0])
-	case yaml.MappingNode:
-		out := map[string]any{}
-		for i := 0; i+1 < len(n.Content); i += 2 {
-			key := n.Content[i]
-			val := n.Content[i+1]
-			if key.Kind != yaml.ScalarNode {
-				continue
-			}
-			out[key.Value] = yamlNodeToValue(val) // last wins
-		}
-		return out
-	case yaml.SequenceNode:
-		out := make([]any, 0, len(n.Content))
-		for _, c := range n.Content {
-			out = append(out, yamlNodeToValue(c))
-		}
-		return out
-	case yaml.ScalarNode:
-		var v any
-		if err := n.Decode(&v); err == nil {
-			return v
-		}
-		return n.Value
-	case yaml.AliasNode:
-		if n.Alias != nil {
-			return yamlNodeToValue(n.Alias)
-		}
-	}
-	return nil
 }
